// src/api/models/get-todo200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * S2I final
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TodoResponse } from './todo-response';

/**
 * 
 * @export
 * @interface GetTodo200Response
 */
export interface GetTodo200Response {
    /**
     * 
     * @type {TodoResponse}
     * @memberof GetTodo200Response
     */
    'data'?: TodoResponse;
    /**
     * 
     * @type {string}
     * @memberof GetTodo200Response
     */
    'message'?: string;
}